+++ src/internal/cleanEntryString.ts
