+++ src/internal/cleanRendererString.ts
