+++ src/internal/getBrowserType.ts
