+++ src/internal/getEntryVersionNumber.ts
